Job queue 그리고 서버 완성

+++ Server/DummyClient/Packet/GenPackets.cs
@@ -6,7 +6,7 @@ using ServerCore;
 
 public enum PacketID
 {
-    C_Chat = 1,
+	C_Chat = 1,
 	S_Chat = 2,
 	
 }
@@ -21,92 +21,88 @@ interface IPacket
 
 class C_Chat : IPacket
 {
-    public string chat;
+	public string chat;
 
-    public ushort Protocol { get { return (ushort)PacketID.C_Chat; } }
+	public ushort Protocol { get { return (ushort)PacketID.C_Chat; } }
 
-    public void Read(ArraySegment<byte> segment)
-    {
-        ushort count = 0;
+	public void Read(ArraySegment<byte> segment)
+	{
+		ushort count = 0;
 
-        ReadOnlySpan<byte> s = new ReadOnlySpan<byte>(segment.Array, segment.Offset, segment.Count);
-        count += sizeof(ushort);
-        count += sizeof(ushort);
-        ushort chatLen = BitConverter.ToUInt16(s.Slice(count, s.Length - count));
+		ReadOnlySpan<byte> s = new ReadOnlySpan<byte>(segment.Array, segment.Offset, segment.Count);
+		count += sizeof(ushort);
+		count += sizeof(ushort);
+		ushort chatLen = BitConverter.ToUInt16(s.Slice(count, s.Length - count));
 		count += sizeof(ushort);
 		this.chat = Encoding.Unicode.GetString(s.Slice(count, chatLen));
 		count += chatLen;
-        
-    }
+	}
 
-    public ArraySegment<byte> Write()
-    {
-        ArraySegment<byte> segment = SendBufferHelper.Open(4096);
-        ushort count = 0; // 지금까지 버퍼에 몇 바이트가 있는지 추적
-        bool success = true;
+	public ArraySegment<byte> Write()
+	{
+		ArraySegment<byte> segment = SendBufferHelper.Open(4096);
+		ushort count = 0;
+		bool success = true;
 
-        Span<byte> s = new Span<byte>(segment.Array, segment.Offset, segment.Count);
+		Span<byte> s = new Span<byte>(segment.Array, segment.Offset, segment.Count);
 
-        count += sizeof(ushort);
-        success &= BitConverter.TryWriteBytes(s.Slice(count, s.Length - count), (ushort)PacketID.C_Chat);
-        count += sizeof(ushort);
-        ushort chatLen = (ushort)Encoding.Unicode.GetBytes(this.chat, 0, this.chat.Length, segment.Array, segment.Offset + count + sizeof(ushort));
+		count += sizeof(ushort);
+		success &= BitConverter.TryWriteBytes(s.Slice(count, s.Length - count), (ushort)PacketID.C_Chat);
+		count += sizeof(ushort);
+		ushort chatLen = (ushort)Encoding.Unicode.GetBytes(this.chat, 0, this.chat.Length, segment.Array, segment.Offset + count + sizeof(ushort));
 		success &= BitConverter.TryWriteBytes(s.Slice(count, s.Length - count), chatLen);
 		count += sizeof(ushort);
 		count += chatLen;
-        // size
-        success &= BitConverter.TryWriteBytes(s, count);
-        if (success == false)
-            return null;
-        return SendBufferHelper.Close(count);
-    }
-} 
+		success &= BitConverter.TryWriteBytes(s, count);
+		if (success == false)
+			return null;
+		return SendBufferHelper.Close(count);
+	}
+}
 
 class S_Chat : IPacket
 {
-    public int playerId;
+	public int playerId;
 	public string chat;
 
-    public ushort Protocol { get { return (ushort)PacketID.S_Chat; } }
+	public ushort Protocol { get { return (ushort)PacketID.S_Chat; } }
 
-    public void Read(ArraySegment<byte> segment)
-    {
-        ushort count = 0;
+	public void Read(ArraySegment<byte> segment)
+	{
+		ushort count = 0;
 
-        ReadOnlySpan<byte> s = new ReadOnlySpan<byte>(segment.Array, segment.Offset, segment.Count);
-        count += sizeof(ushort);
-        count += sizeof(ushort);
-        this.playerId = (int)BitConverter.ToUInt32(s.Slice(count, s.Length - count));
+		ReadOnlySpan<byte> s = new ReadOnlySpan<byte>(segment.Array, segment.Offset, segment.Count);
+		count += sizeof(ushort);
+		count += sizeof(ushort);
+		this.playerId = BitConverter.ToInt32(s.Slice(count, s.Length - count));
 		count += sizeof(int);
 		ushort chatLen = BitConverter.ToUInt16(s.Slice(count, s.Length - count));
 		count += sizeof(ushort);
 		this.chat = Encoding.Unicode.GetString(s.Slice(count, chatLen));
 		count += chatLen;
-        
-    }
+	}
 
-    public ArraySegment<byte> Write()
-    {
-        ArraySegment<byte> segment = SendBufferHelper.Open(4096);
-        ushort count = 0; // 지금까지 버퍼에 몇 바이트가 있는지 추적
-        bool success = true;
+	public ArraySegment<byte> Write()
+	{
+		ArraySegment<byte> segment = SendBufferHelper.Open(4096);
+		ushort count = 0;
+		bool success = true;
 
-        Span<byte> s = new Span<byte>(segment.Array, segment.Offset, segment.Count);
+		Span<byte> s = new Span<byte>(segment.Array, segment.Offset, segment.Count);
 
-        count += sizeof(ushort);
-        success &= BitConverter.TryWriteBytes(s.Slice(count, s.Length - count), (ushort)PacketID.S_Chat);
-        count += sizeof(ushort);
-        success &= BitConverter.TryWriteBytes(s.Slice(count, s.Length - count), this.playerId);
+		count += sizeof(ushort);
+		success &= BitConverter.TryWriteBytes(s.Slice(count, s.Length - count), (ushort)PacketID.S_Chat);
+		count += sizeof(ushort);
+		success &= BitConverter.TryWriteBytes(s.Slice(count, s.Length - count), this.playerId);
 		count += sizeof(int);
 		ushort chatLen = (ushort)Encoding.Unicode.GetBytes(this.chat, 0, this.chat.Length, segment.Array, segment.Offset + count + sizeof(ushort));
 		success &= BitConverter.TryWriteBytes(s.Slice(count, s.Length - count), chatLen);
 		count += sizeof(ushort);
 		count += chatLen;
-        // size
-        success &= BitConverter.TryWriteBytes(s, count);
-        if (success == false)
-            return null;
-        return SendBufferHelper.Close(count);
-    }
-} 
+		success &= BitConverter.TryWriteBytes(s, count);
+		if (success == false)
+			return null;
+		return SendBufferHelper.Close(count);
+	}
+}
